refactor: make configuration panel be composed of differ jsx components
